Test _FORTIFY_SOURCE=3
I have added a proposal in glibc to implement a new fortification
level _FORTIFY_SOURCE=3 which enables potentially more expensive
hardening, as is done by __builtin_dynamic_object_size.  This patch
adds support in tests for allocating the buffer dynamically and then
testing _FORTIFY_SOURCE=3 with them.

The current state is that with upstream clang and gcc, all of the
tests fail with _FORTIFY_SOURCE=3.  With the glibc changes to add
_FORTIFY_SOURCE=3, clang passes all tests because it has
__builtin_dynamic_object_size.